Move drainage checks to the async command runner
